+++ src/ray_hits_utils.c
@@ -1,0 +1,35 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ray_hits_utils.c                                   :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: bhowe <bhowe@student.42singapore.sg>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/10/10 14:50:46 by bhowe             #+#    #+#             */
+/*   Updated: 2024/10/10 14:50:53 by bhowe            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minirt.h"
+
+t_vec	get_hitpoint(t_vec origin, t_ray *ray, float t)
+{
+	return (vector_Add(origin, vector_Multiply(ray->vector, t)));
+}
+
+bool	do_quadratic(t_qdtc *qd, float *t)
+{
+	qd->discrim = qd->b * qd->b - 4 * qd->a * qd->c;
+	if (qd->discrim < 0)
+		return (false);
+	qd->discrim_sqrt = sqrt(qd->discrim);
+	qd->t1 = (-qd->b - qd->discrim_sqrt) / (2.0 * qd->a);
+	qd->t2 = (-qd->b + qd->discrim_sqrt) / (2.0 * qd->a);
+	if (qd->t1 > EPSILON)
+		*t = qd->t1;
+	else if (qd->t2 > EPSILON)
+		*t = qd->t2;
+	else
+		return (false);
+	return (true);
+}
